Code_Competence_2/requirements.txt
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
pyarrow==16.0.0
fastparquet==2024.2.0
google-cloud-storage==2.16.0